pyproject.toml
[project]
name = "taskflow"
version = "0.1.0"
description = "REST API платформа для управления проектами и задачами"
requires-python = ">=3.12"
dependencies = [
    "Django>=5.1,<5.2",
    "djangorestframework>=3.15,<3.16",
    "djangorestframework-simplejwt>=5.3,<6.0",
    "django-filter>=24.3,<25.0",
    "django-cors-headers>=4.4,<5.0",
    "drf-spectacular>=0.27,<0.28",
    "psycopg[binary]>=3.2,<4.0",
    "redis>=5.0,<6.0",
    "celery>=5.4,<5.5",
    "channels>=4.1,<5.0",
    "channels-redis>=4.2,<5.0",
    "daphne>=4.1,<5.0",
    "Pillow>=10.4,<11.0",
    "python-dotenv>=1.0,<2.0",
    "gunicorn>=22.0,<23.0",
    "uvicorn[standard]>=0.30,<1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9.0",
    "pytest-django>=4.8,<5.0",
    "pytest-cov>=5.0,<6.0",
    "pytest-asyncio>=0.23,<1.0",
    "factory-boy>=3.3,<4.0",
    "freezegun>=1.5,<2.0",
    "ipython>=8.26,<9.0",
    "django-extensions>=3.2,<4.0",
    "black>=24.0,<25.0",
    "ruff>=0.6,<1.0",
]
prod = [
    "django-storages[s3]>=1.14,<2.0",
    "sentry-sdk>=2.0,<3.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["apps*", "config*", "core*"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["test_*.py"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
source = ["apps"]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/__init__.py",
    "*/admin.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | migrations
)/
'''

[tool.ruff]
line-length = 100
target-version = "py312"
exclude = [
    ".git",
    ".venv",
    "migrations",
    "__pycache__",
]

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "B",
    "C4",
    "UP",
]
ignore = [
    "E501",
    "B008",
    "W191",
    "B904",
    "UP042",
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"config/settings/*.py" = ["F403", "F405"]
"config/asgi.py" = ["E402"]
